Fix pellet tests (#1)
* Run protege in http mode

* Start protege in tests

Start protege and also enable logging

* Update server configuration for protege

* imports cleanup

* unused imports

* remove all unused code in TestUtilities

* fix server tests

* Less fragile scheme comparison

The previous code fails on hostUri.getScheme() == null while this revision
doesn't

* Use more of the correct client where appropriate

* dead import cleanup

* Argument check for getHost

Otherwise we get a silent error down the road

* hard code the null host to localhost for now

* improve error message

* Quick hack to make the server start properly

* set pellet host for config

* Revert "Quick hack to make the server start properly"

This reverts commit 02f9625a2290f6bd8fb90c708e3766a4f576c045.

* gitignore {client,server}/root

* add missing @Override

* Revert "gitignore {client,server}/root"

This reverts commit 408bd5273d0c4a5cb34d08e0535ecab80550f74c.

* Fix config to make all tests pass

* Fetch resources relative to classes

This will make the tests pass when run separately in the IDE as well as maven

* Use recent version of owlapi

* better error message when dir doesn't exist

* Don't hardcode http

* Automatically inline a few methods in classified tests

These methods were just unnecessary indirection

* Remove extra http:// from tests

* Fix the seed

Rather than choosing it randomly. It looks like the set of tests that fails
depends on what the seed is set to. So let's keep it constant for now

* Change some tests to display full results

Previously, tests would only display the first failure and exit