Refactor Pellet Server Implementation (#8)
* Remove pointless private method and imports

* Make ServerStateImpl abstract

This makes it easier to verify that concrete versions of this class aren't used
anywhere. Hence we can simply inline it.

* Remove unused method

* Inline update into ProtegeServerState

This is the 1 method intellij can't figure out how to eliminate because it's
being called. So we can just make it abstract for now and move the
implementation ourselves.

* Remove mHome instance variable

It was never used so should just be removed

* Inline ServerStateImpl into ProtegeServer

This parent class serves no purpose. Its implementation isn't used anywhere
apart from the subclass and any abstraction boundary it might provide is already
provided by the interface. Hence it's implementation is moved to the only place
where it's used - the subclass.