Fetch classifiable projects from protege (#20)
* Crate Pellet Auth Handler

This will wrap pellet's restart/shutdown handlers. And anything else
that might need authentication.

* Add an interceptor to add authorization to pellet's client calls

The interceptor will the add password it's initialized with. The username
stays constant for simplicity.

* Initialize pellet client with auth password

if it's provided in the settings

* add accessor for fetching the password from config

* Make restart/shutdown require credentials

* Update tests to use auth stuff

* Pass management password in the other constructor

when it's available

* remove pointless method

* Abstract over fetching ontologies

Because the tests really rely on having a static list of ontologies, we need
a way to maintain that. The approach taken is to have an ontology provider
that will have 2 implementations depending on what we want the source of the
ontologies to be.

* Fetch ontologies from protege

Add a new implementation for the ontology fetcher. This implementation will use
the classifiable endpoint in protege to fetch the classifiable project id's.

* remove a little bit of dead code

* Use guava Optional when necessary

* missing @Override

* Use guava optional in injector